enhanced version of SqlPredicateTest (first draft)
